Splitted timeline series into several files.
